Prevent destroying the "anonymous" token on Consul-pre 1.4
It's not possible to destroy this token at all as it's special.
This merely resets it to its default value and removes it from the
Terraform state so it's not tracked anymore.

This should help Terraform configuration running against old versions of
Consul, which tries to reconfigure this anonymous token, while preventing
error when that state is destroyed.